requirements.txt
streamlit==1.22.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
plotly==5.18.0
pyyaml==6.0
joblib==1.3.2
twisted==22.10.0  # For OPC UA simulation